Implement local development
