+++ Procfile
@@ -1,2 +1,2 @@
-web: gunicorn wsgi:app
+web: gunicorn web:app
 worker: celery -A tasks.celery worker -c 4
