+++ config.py
@@ -13,17 +13,15 @@ from pathlib import Path
 #   russian.json  # russian.xml
 
 APP_NAME = "witty-bot"
+APP_BASE_URL = os.getenv('APP_URL', "http://localhost:8080")
 
-if os.getenv("DYNO"):
-    APP_BASE_URL = f"https://{APP_NAME}.herokuapp.com"
-else:
-    APP_BASE_URL = "http://localhost:8080"
+MEDIA_URL = '/media'
+MEDIA_PATH = Path("/tmp")
 
 PROJECT_DIR = os.path.dirname(os.path.abspath(__file__))
 RESOURCES_PATH = Path(PROJECT_DIR, "resources")
 MATH_PROBLEMS_PATH = RESOURCES_PATH / "math"
 BOOKS_PATH = RESOURCES_PATH / "books"
-MEDIA_PATH = Path("/tmp")
 
 TELEGRAM_BOT_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN")
 GOOGLE_SERVICE_ACCOUNT_INFO = os.getenv("GOOGLE_SERVICE_ACCOUNT_INFO")
